tjBufSize(): Fix int overflow w/ large dimensions
The intermediate PAD(width, mcuw) * PAD(height, mcuh) computation was
performed using signed integer arithmetic and could thus overflow if
PAD(width, mcuw) * PAD(height, mcuh) > INT_MAX.  This was innocuous,
however, because the API requires that the allocated buffer size be
reported to tjCompress*() or tjTransform().  The calling program would
presumably pass the (incorrect) return value from tjBufSize() to
tjAlloc() and report the same size to tjCompress*() or tjTransform().
Depending on whether TJFLAG_NOREALLOC was set,
tjCompress*()/tjTransform() would either reallocate the buffer as needed
or fail if the reported buffer size was exceeded.  As long as the
reported buffer size was the same as the allocated buffer size, no
buffer overrun could occur.  Thus, the only issue was a usability issue:
passing large dimensions to tjBufSize(), then passing the return value
of tjBufSize() to tjAlloc() and tjCompress*()/tjTransform(), could have
caused an error ("Buffer passed to JPEG library is too small") if
TJFLAG_NOREALLOC was set.

+++ turbojpeg.c
@@ -555,7 +555,8 @@ DLLEXPORT unsigned long tjBufSize(int width, int height, int jpegSubsamp)
   mcuw = tjMCUWidth[jpegSubsamp];
   mcuh = tjMCUHeight[jpegSubsamp];
   chromasf = jpegSubsamp == TJSAMP_GRAY ? 0 : 4 * 64 / (mcuw * mcuh);
-  retval = PAD(width, mcuw) * PAD(height, mcuh) * (2ULL + chromasf) + 2048ULL;
+  retval = (unsigned long long)PAD(width, mcuw) * PAD(height, mcuh) *
+           (2ULL + chromasf) + 2048ULL;
 #if ULLONG_MAX > ULONG_MAX
   if (retval > (unsigned long long)((unsigned long)-1))
     THROWG("tjBufSize(): Image is too large");
